Updated EbotsmotionController to function with the state machine auton opmode by removing internal loops and promoting loop variables to class attributes.
Corrected glitch in SoftStart scale calculation.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SoftStart.java
@@ -36,7 +36,7 @@ public enum SoftStart {
 
         double scaleFactor = 1.0;
         if(currentTimeMillis < durationMillis){
-            scaleFactor = (double) (currentTimeMillis/durationMillis);
+            scaleFactor = ((double)currentTimeMillis/ (double) durationMillis);
             if (scaleFactor<minPower) scaleFactor=minPower;
         }
